move matmul tests to separate file
