+++ tests/python/test_python_frontend.py
@@ -3,7 +3,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Owner(s): ["module: nvfuser"]
 
-from copy import deepcopy
 from functools import partial
 import itertools
 import math
@@ -2300,98 +2299,6 @@ class TestNvFuserFrontend(TestCase):
 
         self.assertTrue(nvf_out[0].device.index == 1)
 
-    def test_matmul(self):
-        m = 24
-        n = 16
-        k = 8
-        inputs_tt = [
-            torch.randn(m, k, device="cuda", dtype=torch.float16),
-            torch.randn(k, n, device="cuda", dtype=torch.float16),
-        ]
-
-        inputs_tn = [
-            inputs_tt[0].clone(),
-            inputs_tt[1].clone().as_strided(size=[k, n], stride=[1, k]),
-        ]
-
-        inputs_nt = [
-            inputs_tt[0].clone().as_strided(size=[m, k], stride=[1, m]),
-            inputs_tt[1].clone(),
-        ]
-
-        inputs_tn = [inputs_tt[0].clone(), inputs_tn[1].clone()]
-
-        inputs_nn = [inputs_nt[0].clone(), inputs_tn[1].clone()]
-
-        def fusion_func(fd: FusionDefinition, inps) -> None:
-            t0 = fd.from_pytorch(inps[0])
-            t1 = fd.from_pytorch(inps[1])
-            t2 = fd.ops.matmul(t0, t1)
-            fd.add_output(t2)
-
-        for inps in [inputs_tt, inputs_tn, inputs_nt, inputs_nn]:
-            nvf_out, _ = exec_nvfuser(partial(fusion_func, inps=inps), inps)
-            eager_out = torch.matmul(inps[0], inps[1])
-            fp16_nvf_out = nvf_out[0]
-            self.assertEqual(eager_out, fp16_nvf_out)
-
-    def test_linear(self):
-        m = 24
-        n = 16
-        k = 8
-        bias0d = torch.tensor(3.14, device="cuda", dtype=torch.float16)
-        bias1d = torch.randn(n, device="cuda", dtype=torch.float16)
-
-        inputs_mk_nk = [
-            torch.randn(m, k, device="cuda", dtype=torch.float16),
-            torch.randn(n, k, device="cuda", dtype=torch.float16),
-        ]
-
-        inputs_mk_kn = [
-            inputs_mk_nk[0].clone(),
-            inputs_mk_nk[1].clone().as_strided(size=[n, k], stride=[1, n]),
-        ]
-
-        inputs_km_nk = [
-            inputs_mk_nk[0].clone().as_strided(size=[m, k], stride=[1, m]),
-            inputs_mk_nk[1].clone(),
-        ]
-
-        inputs_km_kn = [
-            inputs_km_nk[0].clone(),
-            inputs_mk_kn[1].clone(),
-        ]
-
-        def fusion_func(
-            fd: FusionDefinition,
-            inp: torch.Tensor,
-            wt: torch.Tensor,
-            bias: torch.Tensor | None,
-        ) -> None:
-            t0 = fd.from_pytorch(inp)
-            t1 = fd.from_pytorch(wt)
-            if bias is not None:
-                t2 = fd.from_pytorch(bias)
-                t_out = fd.ops.linear(t0, t1, t2)
-            else:
-                t_out = fd.ops.linear(t0, t1)
-            fd.add_output(t_out)
-
-        in_tensors = [inputs_mk_nk, inputs_mk_kn, inputs_km_nk, inputs_km_kn]
-        use_bias = [None, bias0d, bias1d]
-        for [inp, wt], use_bias in list(itertools.product(in_tensors, use_bias)):
-            with self.subTest(inp=inp, wt=wt, use_bias=use_bias):
-                input_tensors = (
-                    (inp, wt, use_bias) if use_bias is not None else (inp, wt)
-                )
-                nvf_out, _ = exec_nvfuser(
-                    partial(fusion_func, inp=inp, wt=wt, bias=use_bias),
-                    input_tensors,
-                )
-                eager_out = F.linear(input=inp, weight=wt, bias=use_bias)
-                fp16_nvf_out = nvf_out[0]
-                torch.testing.assert_close(fp16_nvf_out, eager_out, atol=1e-3, rtol=0)
-
     def test_integer_division(self):
         inputs = [
             torch.testing.make_tensor(1024, device="cuda", dtype=torch.long),
@@ -4257,43 +4164,6 @@ class TestNvFuserFrontend(TestCase):
                 assert (
                     match_pairs.item() < 3
                 ), f"At least three entries match in {output}"
-
-    def test_matmul_issue_2354(self):
-        inputs = [
-            torch.randn((8, 4), dtype=torch.float32, device="cuda:0"),
-            torch.randn(
-                (
-                    6,
-                    2,
-                    4,
-                ),
-                dtype=torch.float32,
-                device="cuda:0",
-            ),
-        ]
-
-        def fusion_func(fd: FusionDefinition):
-            T0 = fd.define_tensor(
-                shape=[-1, -1],
-                contiguity=[True, True],
-                dtype=DataType.Float,
-                is_cpu=False,
-                stride_order=[1, 0],
-            )
-            T1 = fd.define_tensor(
-                shape=[-1, -1, -1],
-                contiguity=[True, True, True],
-                dtype=DataType.Float,
-                is_cpu=False,
-                stride_order=[2, 1, 0],
-            )
-            T2 = fd.ops.linear(T1, T0)
-            S3 = fd.define_scalar(1.41421, dtype=DataType.Double)
-            T4 = fd.ops.mul(T2, S3)
-            fd.add_output(T2)
-            fd.add_output(T4)
-
-        nvf_out, _ = exec_nvfuser(fusion_func, inputs)
 
     def test_reshape_dynamic(self):
         inputs = [
